+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * New logger `TagLogger` and now extension `Any#logger` will use it to decrease objects allocations
 * Now it is possible to create `KSLog` using any iterable
+* Now it is possible to create `KSLog` using vararg log levels
 
 ## 0.0.1